Add promo check utility
